refactor: move DefinitionAdapter to new file
