+++ src/language/common/lspLanguageFeatures.ts
@@ -13,6 +13,7 @@ import {
 	CancellationToken
 } from '../../fillers/monaco-editor-core';
 import { fromPosition, toRange, toTextEdit, fromRange } from './CompletionAdapter';
+import { toLocation } from './DefinitionAdapter';
 
 export interface WorkerAccessor<T> {
 	(...more: Uri[]): Promise<T>;
@@ -26,45 +27,7 @@ export * from './HoverAdapter';
 
 export * from './DocumentHighlightAdapter';
 
-//#region DefinitionAdapter
-
-export interface ILanguageWorkerWithDefinitions {
-	findDefinition(uri: string, position: lsTypes.Position): Promise<lsTypes.Location | null>;
-}
-
-export class DefinitionAdapter<T extends ILanguageWorkerWithDefinitions>
-	implements languages.DefinitionProvider
-{
-	constructor(private readonly _worker: WorkerAccessor<T>) {}
-
-	public provideDefinition(
-		model: editor.IReadOnlyModel,
-		position: Position,
-		token: CancellationToken
-	): Promise<languages.Definition | undefined> {
-		const resource = model.uri;
-
-		return this._worker(resource)
-			.then((worker) => {
-				return worker.findDefinition(resource.toString(), fromPosition(position));
-			})
-			.then((definition) => {
-				if (!definition) {
-					return;
-				}
-				return [toLocation(definition)];
-			});
-	}
-}
-
-function toLocation(location: lsTypes.Location): languages.Location {
-	return {
-		uri: Uri.parse(location.uri),
-		range: toRange(location.range)
-	};
-}
-
-//#endregion
+export * from './DefinitionAdapter';
 
 //#region ReferenceAdapter
